Changed where the options to add admins / locations are
As per peer testing
